weird changes?
You are not necessarily using importing query, so I've removed it. I changed the "any" type from the code and password variables you placed and added this "toString()" method I read somewhere to help convert all query parameters to strings; that way, there won't be type errors (prolly). Also, I used "toString()" method again, but this time to replace the "where" class in the prisma query to convert the code parameter to a string to match the data type in the prisma schema. But bruhhh, I just realised today it was Prisma. I didn't check at all. i had no idea the code was public.